requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
pillow==12.2.0
python-dotenv==1.2.2
six==1.17.0
sqlparse==0.5.5
